# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: S.A.L.I.M. Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /objectives/{id}/specify:
    get:
      operationId: getSpecify
      tags: [specify]
      summary: Get Specify data for an objective
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Specify data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecifyData"
    put:
      operationId: saveSpecify
      tags: [specify]
      summary: Save Specify step data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecifyInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecifyData"

  # ─── A – ALIGN ──────────────────────────────────────────
  /objectives/{id}/align:
    get:
      operationId: getAlign
      tags: [align]
      summary: Get Align data for an objective
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Align data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlignData"
    put:
      operationId: saveAlign
      tags: [align]
      summary: Save Align step data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlignInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlignData"

  # ─── L – LAY OUT ────────────────────────────────────────
  /objectives/{id}/plan:
    get:
      operationId: getPlan
      tags: [layout]
      summary: Get the 90-day plan for an objective
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan with milestones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    put:
      operationId: savePlan
      tags: [layout]
      summary: Save the 90-day plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "200":
          description: Saved plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /objectives/{id}/plan/milestones:
    post:
      operationId: createMilestone
      tags: [layout]
      summary: Add a milestone to the plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [layout]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [layout]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── I – IMPLEMENT (TASKS) ──────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (optionally filtered by date)
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
          description: ISO date string (YYYY-MM-DD) to filter tasks for a specific day
        - name: objectiveId
          in: query
          schema:
            type: ["integer", "null"]
          description: Filter by objective
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new daily task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # ─── I – IMPLEMENT (WEEKLY REVIEWS) ─────────────────────
...
components:
  schemas:
...
    SpecifyData:
      type: object
      required: [id, objectiveId]
      properties:
        id:
          type: integer
        objectiveId:
          type: integer
        smartGoal:
          type: ["string", "null"]
        specificWhat:
          type: ["string", "null"]
        measurableHow:
          type: ["string", "null"]
        achievableSteps:
          type: ["string", "null"]
        relevantWhy:
          type: ["string", "null"]
        timeBoundWhen:
          type: ["string", "null"]
        precisionScore:
          type: ["integer", "null"]
        isValidated:
          type: boolean
    SpecifyInput:
      type: object
      properties:
        smartGoal:
          type: string
        specificWhat:
          type: string
        measurableHow:
          type: string
        achievableSteps:
          type: string
        relevantWhy:
          type: string
        timeBoundWhen:
          type: string
        precisionScore:
          type: integer
        isValidated:
          type: boolean

    # ─── ALIGN ─────────────────────────────────────────────
    AlignData:
      type: object
      required: [id, objectiveId]
      properties:
        id:
          type: integer
        objectiveId:
          type: integer
        deepWhy:
          type: ["string", "null"]
        values:
          type: array
          items:
            type: string
        visualizationNotes:
          type: ["string", "null"]
        alignmentScore:
          type: ["integer", "null"]
        isCompleted:
          type: boolean
    AlignInput:
      type: object
      properties:
        deepWhy:
          type: string
        values:
          type: array
          items:
            type: string
        visualizationNotes:
          type: string
        alignmentScore:
          type: integer
        isCompleted:
          type: boolean

    # ─── PLAN (LAY OUT) ────────────────────────────────────
    Plan:
      type: object
      required: [id, objectiveId, milestones]
      properties:
        id:
          type: integer
        objectiveId:
          type: integer
        summary:
          type: ["string", "null"]
        priorityFocus:
          type: ["string", "null"]
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
    PlanInput:
      type: object
      properties:
        summary:
          type: string
        priorityFocus:
          type: string
    Milestone:
      type: object
      required: [id, planId, title, weekNumber, isCompleted]
      properties:
        id:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        weekNumber:
          type: integer
        targetDate:
          type: ["string", "null"]
          format: date
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
    MilestoneInput:
      type: object
      required: [title, weekNumber]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        weekNumber:
          type: integer
        targetDate:
          type: string
          format: date
    MilestoneUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        weekNumber:
          type: integer
        targetDate:
          type: string
          format: date
        isCompleted:
          type: boolean

    # ─── TASKS ─────────────────────────────────────────────
    Task:
      type: object
      required: [id, userId, title, isCompleted, scheduledDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        objectiveId:
          type: ["integer", "null"]
        milestoneId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        scheduledDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title, scheduledDate]
      properties:
        objectiveId:
          type: integer
        milestoneId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        scheduledDate:
          type: string
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        isCompleted:
          type: boolean
        scheduledDate:
          type: string
          format: date

    # ─── WEEKLY REVIEWS ────────────────────────────────────
...